pyproject.toml
[tool.poetry]
name = "sso"
version = "0.1.0"
description = "SSO application"
authors = ["Irtaza Hussain"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
pymongo = {extras = ["srv"], version = "^4.6.0"}
pylint = "^3.0.2"
pre-commit = "^3.5.0"
flake8 = "^6.1.0"
python-jose = "^3.3.0"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
